+++ Libraries/LibWeb/Painting/BorderRadiusCornerClipper.cpp
@@ -4,19 +4,20 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <LibWeb/Painting/BorderRadiiData.h>
 #include <LibWeb/Painting/BorderRadiusCornerClipper.h>
 #include <LibWeb/Painting/DisplayListRecorder.h>
 #include <LibWeb/Painting/DisplayListRecordingContext.h>
 
 namespace Web::Painting {
 
-ScopedCornerRadiusClip::ScopedCornerRadiusClip(DisplayListRecordingContext& context, DevicePixelRect const& border_rect, BorderRadiiData const& border_radii, CornerClip corner_clip, bool do_apply)
+ScopedCornerRadiusClip::ScopedCornerRadiusClip(DisplayListRecordingContext& context, DevicePixelRect const& border_rect, BorderRadiiData const& border_radii, Gfx::CornerClip corner_clip, bool do_apply)
     : m_context(context)
 {
     m_do_apply = do_apply;
     if (!do_apply)
         return;
-    CornerRadii const corner_radii {
+    Gfx::CornerRadii const corner_radii {
         .top_left = border_radii.top_left.as_corner(context.device_pixel_converter()),
         .top_right = border_radii.top_right.as_corner(context.device_pixel_converter()),
         .bottom_right = border_radii.bottom_right.as_corner(context.device_pixel_converter()),
