+++ Libraries/LibWeb/Painting/CanvasPaintable.cpp
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <LibWeb/Painting/BorderRadiusCornerClipper.h>
 #include <LibWeb/Painting/CanvasPaintable.h>
 #include <LibWeb/Painting/DisplayListRecorder.h>
